# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "This is a simple bookstore server with a book inventory, users and\
    \ a shopping cart."
  version: "1.0.0"
  title: "Simple Bookstore"
  contact:
    email: "matteo.pozzi0397@gmail.com"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/v2"
tags:
- name: "book"
  description: "Available book"
- name: "genre"
  description: "Literary genre"
- name: "theme"
  description: "Literary theme"
- name: "author"
  description: "Author of a book"
- name: "event"
  description: "Event to present a book"
- name: "cart"
  description: "Access to the cart"
- name: "user"
  description: "Operations about user"
schemes:
- "http"
paths:
  /books:
    get:
      tags:
      - "book"
      summary: "Books available in the inventory"
      description: "List of books available in the inventory"
      operationId: "booksGET"
      produces:
      - "application/json"
      parameters:
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      - name: "author"
        in: "query"
        description: "Filter list of books by author"
        required: false
        type: "integer"
      - name: "genre"
        in: "query"
        description: "Filter list of books about the same genre"
        required: false
        type: "string"
      - name: "theme"
        in: "query"
        description: "Filter list of books which treat the same theme"
        required: false
        type: "string"
      responses:
        200:
          description: "A collection of Books"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Book"
        404:
          description: "Unexpected error"
      x-swagger-router-controller: "Book"
  /books/{bookId}:
    get:
      tags:
      - "book"
      summary: "Find book by ID"
      description: "Returns a book"
      operationId: "getBookById"
      produces:
      - "application/json"
      parameters:
      - name: "bookId"
        in: "path"
        description: "ID of book to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Book"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Book not found"
      x-swagger-router-controller: "Book"
  /books/genres:
    get:
      tags:
      - "genre"
      summary: "Literary genres of books"
      description: "List of literary genres the books belong to"
      operationId: "genresGET"
      produces:
      - "application/json"
      parameters:
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "A collection of Genres"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Genre"
        404:
          description: "Unexpected error"
      x-swagger-router-controller: "Genre"
  /books/themes:
    get:
      tags:
      - "theme"
      summary: "Themes treated by books"
      description: "List of themes the books treat"
      operationId: "themesGET"
      produces:
      - "application/json"
      parameters:
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "A collection of Themes"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Theme"
        404:
          description: "Unexpected error"
      x-swagger-router-controller: "Theme"
  /user/login:
    post:
      tags:
      - "user"
      summary: "Login"
      description: "Login with a form"
      operationId: "userLoginPOST"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "formData"
        required: true
        type: "string"
      - name: "password"
        in: "formData"
        required: true
        type: "string"
      responses:
        200:
          description: "successful login"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "User"
  /user/register:
    post:
      tags:
      - "user"
      summary: "Register"
      description: "Register into the store"
      operationId: "userRegisterPOST"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "successful registration"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "User"
  /cart/{cartId}:
    get:
      tags:
      - "cart"
      summary: "View the content of the cart"
      operationId: "cartCartIdGET"
      produces:
      - "application/json"
      parameters:
      - name: "cartId"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Cart"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "Cart"
  /authors:
    get:
      tags:
      - "author"
      summary: "Authors who wrote a book available in the store"
      description: "List of authors of books available in the inventory"
      operationId: "authorsGET"
      produces:
      - "application/json"
      parameters:
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      - name: "book"
        in: "query"
        description: "Filter authors who wrote the given book."
        required: false
        type: "integer"
      responses:
        200:
          description: "A collection of Authors"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Author"
        404:
          description: "Unexpected error"
      x-swagger-router-controller: "Author"
  /authors/{authorId}:
    get:
      tags:
      - "author"
      summary: "Find author by ID"
      description: "Returns an author"
      operationId: "getAuthorById"
      produces:
      - "application/json"
      parameters:
      - name: "authorId"
        in: "path"
        description: "ID of author to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Author"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Author not found"
      x-swagger-router-controller: "Author"
  /events:
    get:
      tags:
      - "event"
      summary: "Events organized by the store"
      description: "List of events organized by the store"
      operationId: "eventsGET"
      produces:
      - "application/json"
      parameters:
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      - name: "book"
        in: "query"
        description: "Filter events where given book is presented"
        required: false
        type: "integer"
      - name: "author"
        in: "query"
        description: "Filter events presented by an author"
        required: false
        type: "integer"
      responses:
        200:
          description: "A collection of Events"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Event"
        404:
          description: "Unexpected error"
      x-swagger-router-controller: "Event"
  /events/{eventId}:
    get:
      tags:
      - "event"
      summary: "Find event by ID"
      description: "Returns an event"
      operationId: "getEventById"
      produces:
      - "application/json"
      parameters:
      - name: "eventId"
        in: "path"
        description: "ID of event to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Event"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Event not found"
      x-swagger-router-controller: "Event"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      address:
        type: "string"
      creditcard:
        type: "string"
    title: "User"
    description: "A user"
    example:
      id: 1
      name: "Matteo"
      address: "PoliMi"
      creditcard: "xyzabc"
  Cart:
    type: "object"
    properties:
      total:
        $ref: "#/definitions/Amount"
      books:
        type: "array"
        items:
          $ref: "#/definitions/Book"
    title: "Cart"
    description: "Order for books"
    example:
      total:
        currency: "eur"
        value: 80082819046101.171875
      books:
      - id: 0
        title: "Il deserto dei tartari"
        author: "Dino Buzzati"
        price:
          value: 10
          currency: "eur"
      - id: 0
        title: "Il deserto dei tartari"
        author: "Dino Buzzati"
        price:
          value: 10
          currency: "eur"
  Book:
    type: "object"
    required:
    - "price"
    - "title"
    properties:
      id:
        type: "integer"
        format: "int64"
      title:
        type: "string"
      authors:
        type: "array"
        uniqueItems: true
        items:
          $ref: "#/definitions/Author"
        maxItems: 4
        minItems: 1
      genres:
        type: "array"
        uniqueItems: true
        items:
          $ref: "#/definitions/Genre"
        maxItems: 3
        minItems: 1
      themes:
        type: "array"
        uniqueItems: true
        items:
          $ref: "#/definitions/Theme"
        maxItems: 6
        minItems: 1
      price:
        $ref: "#/definitions/Amount"
      copies:
        type: "integer"
        description: "number of copies available of the book"
        minimum: 0
        maximum: 20
      status:
        type: "string"
        description: "book availability in the inventory"
        enum:
        - "available"
        - "out of stock"
    title: "Book"
    description: "A book for sale in the store"
    example:
      id: 0
      title: "Il deserto dei tartari"
      authors:
      - "Dino Buzzati"
      genres:
      - "fiction"
      themes:
      - "self-deception"
      copies: 3
      price:
        value: 10
        currency: "eur"
      status: "available"
  Author:
    type: "object"
    required:
    - "name"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    title: "Author"
    description: "An author who wrote a book in the store"
    example:
      id: 0
      name: "Dino Buzzati"
  Genre:
    type: "string"
    description: "A genre of a book"
    enum:
    - "fantasy"
    - "science fiction"
    - "adventure"
    - "action"
    - "biography"
    - "autobiography"
    - "thriller"
    - "romance"
    - "fiction"
    - "horror"
    - "poetry"
    - "drama"
    - "historical fiction"
    - "fairytale"
    - "scientific paper"
  Theme:
    type: "string"
    description: "A theme treated by a book"
    enum:
    - "war"
    - "love"
    - "death"
    - "exile"
    - "travel"
    - "friendship"
    - "family"
    - "immortality"
    - "chaos"
    - "coming of age"
    - "survival"
    - "self-deception"
    - "courage"
    - "good vs evil"
    - "corruption"
    - "racism"
    - "fear of failure"
    - "betrayal"
  Event:
    type: "object"
    required:
    - "book"
    - "date"
    - "place"
    properties:
      id:
        type: "integer"
        format: "int64"
      title:
        type: "string"
      date:
        type: "string"
        format: "date-time"
      place:
        type: "string"
      book:
        $ref: "#/definitions/Book"
    title: "Event"
    description: "An event in which a book will be presented"
    example:
      id: 0
      title: "Buzzati Grand Tour"
      date: "2019-24-10T10:00:00Z"
      place: "Piazza Leonardo da Vinci 1 Milano"
      book:
        id: 0
        title: "Il deserto dei tartari"
        author: "Dino Buzzati"
        price:
          value: 10
          currency: "eur"
        status: "available"
  Amount:
    type: "object"
    required:
    - "currency"
    - "value"
    properties:
      value:
        type: "number"
        format: "double"
        minimum: 0.01
        maximum: 1000000000000000
      currency:
        $ref: "#/definitions/Currency"
    description: "Price\n"
    example:
      currency: "eur"
      value: 11
  Currency:
    type: "string"
    pattern: "^[A-Z]{3,3}$"
    description: "some description\n"
    example: "eur"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"